% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkedscatter.R
\docType{package}
\name{linkedscatter}
\alias{linkedscatter}
\alias{linkedscatter-package}
\title{linkedscatter-package}
\usage{
linkedscatter(x = NULL, y = NULL, tooltip = NULL, width = NULL,
  height = NULL)
}
\description{
linkedscatter-package

Add a linked scatter plot
}

