% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkedscatter.R
\name{linkedscatterOutput}
\alias{linkedscatterOutput}
\title{Widget output function for use in Shiny}
\usage{
linkedscatterOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

