% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/linkedscatter.R
\name{renderLinkedscatter}
\alias{renderLinkedscatter}
\title{Widget render function for use in Shiny}
\usage{
renderLinkedscatter(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

